+++ transport_test.go
@@ -32,14 +32,14 @@ func TestTransport(t *testing.T) {
 		ConsumerKey:    expectedConsumerKey,
 		ConsumerSecret: "consumer_secret",
 	}
-	signer := &Signer{
+	auther := &auther{
 		config: config,
 		clock:  &fixedClock{time.Unix(123456789, 0)},
 		noncer: &fixedNoncer{expectedNonce},
 	}
 	tr := &Transport{
 		source: StaticTokenSource(NewToken(expectedToken, "some_secret")),
-		signer: signer,
+		auther: auther,
 	}
 	client := &http.Client{Transport: tr}
 
@@ -67,7 +67,7 @@ func TestTransport_customBaseTransport(t *testing.T) {
 func TestTransport_nilSource(t *testing.T) {
 	tr := &Transport{
 		source: nil,
-		signer: &Signer{
+		auther: &auther{
 			config: &Config{},
 			clock:  &fixedClock{time.Unix(123456789, 0)},
 			noncer: &fixedNoncer{"any_nonce"},
@@ -84,7 +84,7 @@ func TestTransport_nilSource(t *testing.T) {
 func TestTransport_emptySource(t *testing.T) {
 	tr := &Transport{
 		source: StaticTokenSource(nil),
-		signer: &Signer{
+		auther: &auther{
 			config: &Config{},
 			clock:  &fixedClock{time.Unix(123456789, 0)},
 			noncer: &fixedNoncer{"any_nonce"},
@@ -98,16 +98,16 @@ func TestTransport_emptySource(t *testing.T) {
 	}
 }
 
-func TestTransport_nilSigner(t *testing.T) {
+func TestTransport_nilAuther(t *testing.T) {
 	tr := &Transport{
 		source: StaticTokenSource(&Token{}),
-		signer: nil,
+		auther: nil,
 	}
 	client := &http.Client{Transport: tr}
 	resp, err := client.Get("http://example.com")
 	assert.Nil(t, resp)
 	if assert.Error(t, err) {
-		assert.Equal(t, "Get http://example.com: oauth1: Transport's signer is nil", err.Error())
+		assert.Equal(t, "Get http://example.com: oauth1: Transport's auther is nil", err.Error())
 	}
 }